requirements.txt
pydantic
langgraph
langchain-core
langchain-ollama
llama-cpp-python
py3Dmol

huggingface_hub
ollama
numpy==1.26.2
scikit-learn==1.1.2

generate-chroma==1.0.0
langchain-community==0.3.20
py3Dmol
omegaconf
fair-esm
openfold
prody